requirements.txt
# E-Waste Detection System - Dependencies

# Core ML / CV
ultralytics>=8.0.0
torch>=2.0.0
torchvision>=0.15.0

# Image Processing
opencv-python>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0

# Data & Utilities
pandas>=2.0.0
PyYAML>=6.0
matplotlib>=3.7.0

# Augmentation
albumentations>=1.3.0

# Demo Interface
streamlit>=1.28.0

# Notebook support
ipykernel>=6.0.0
ipywidgets>=8.0.0

# AI Integration
google-generativeai>=0.3.0
python-dotenv>=1.0.0
